requirements.txt
streamlit
pandas
numpy
Pillow
openpyxl